+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # changelog
 
+## Version 6.1.2
+
+* add --node-gyp-options=<options> param in compilation script [PR11](https://github.com/UrielCh/opencv4nodejs/pull/11)
+* add doc in code
+* rename _binding.gyp to binding.gyp and add a dummy "install" script
+* add missing dev dependency
+* improve compilation output log
+* improve --dry-run mode
+
 ## Version 6.1.1
 
 * fix ambiguous typing